mattermost-master/groupbuy-bot/core/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
redis>=5.0
gunicorn>=21.0
python-dotenv>=1.0
celery>=5.3
django-cors-headers>=4.3
drf-spectacular>=0.27
PyJWT>=2.8
aiohttp>=3.9
